backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-generativeai
requests
trafilatura